+++ tests/e2e/09-table-transfer.spec.js
@@ -1,14 +1,9 @@
 import { test, expect } from "@playwright/test";
 import { startEditor, waitForEditor } from "./helpers.js";
 import {
-  cmdArrow,
   expectNoReload,
   persistedFirstTableRows,
-  rangeBounds,
-  rangeCellTexts,
-  readCellText,
   selectCell,
-  selectedText,
   setReloadMarker,
 } from "./table-helpers.js";